requirements.txt
numpy
pandas
matplotlib
mlxtend
torchmetrics
tqdm
prettytable
requests
zipfile